tulong is always 32 bit, since that's what the code expects. (sizeof is used in some file loading routines)
git-svn-id: svn://svn.code.sf.net/p/therush/code@41 4b14b72a-3063-4cf0-bd6b-70bde676b5ec

+++ sources/libbase/ZBaseDefs.h
@@ -26,8 +26,10 @@
 #include <stdlib.h>
 #endif
 
+#include <stdint.h>
 
-typedef unsigned long tulong;
+
+typedef uint32_t tulong;
 typedef long tlong;
 typedef double tdouble;
